Added framer motion to home page.
