+++ src/transitions.js
@@ -1,0 +1,8 @@
+export const transition1 = {
+  duration: 1.2,
+  // ease: [0.6, 0.01, -0.05, 0.9],
+  ease: [0.43, 0.13, 0.23, 0.96],
+};
+export const transition2 = {
+  duration: 1.8,
+};
