--- src/transitions.jsx
@@ -1,5 +1,0 @@
-export const transition1 = {
-  duration: 1.2,
-  ease: [0.6, 0.01, -0.05, 0.9],
-  // ease: [0.43, 0.13, 0.23, 0.96],
-};
